+++ locale/fr_FR.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * eduVPN - End-user friendly VPN.
+ *
+ * Copyright: 2016-2020, The Commons Conservancy eduVPN Programme
+ * SPDX-License-Identifier: AGPL-3.0+
+ */
+
+return [
+    'Android' => 'Android',
+    'Continue...' => 'Continuer...',
+    'Linux' => 'Linux',
+    'On the "Configurations" page, advanced users can manually download VPN configurations.' => 'Sur la page "Configurations", les utilisateurs avancés peuvent manuellement télécharger une configuration VPN.',
+    'On your "Account" page you can block access to the VPN in case you lose a device.' => 'Sur la page "Compte" vous pouvez bloquer l\'accès au VPN depuis les terminaux que vous auriez perdu.',
+    'To use Let\'s Connect!, download the app for your device below!' => 'Pour utiliser Let\'s Connect!, téléchargez l\'application correspondant à votre terminal ci-dessous !',
+    'Welcome to Let\'s Connect!' => 'Bienvenue sur Let\'s Connect!',
+    'Windows' => 'Windows',
+    'iOS' => 'iOS',
+    'macOS' => 'macOS',
+];