+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 2.2.1 (...)
 - add `de_DE` translation
 - add `pl_PL` translation
+- add `fr_FR` translation
 
 ## 2.2.0 (2020-06-29)
 - implement new design